requirements.txt
tensorflow>=2.10
numpy>=1.21
Pillow>=9.0

